WFCORE-6791 Tidy wildfly-discovery code to better use as a reference for wildfly-subsystem usage
